+++ cli/tests/ui/config_schema_additional_properties_schema.stderr
@@ -1,0 +1,12 @@
+Error: manifest::invalid_config_schema
+
+  × check failed
+  ╰─▶ invalid config definition at additionalProperties:
+      `additionalProperties` must be a boolean in Amber config_schema
+   ╭─[<CARGO_MANIFEST_DIR>/tests/ui/config_schema_additional_properties_schema.json5:8:27]
+ 7 │     },
+ 8 │     additionalProperties: { type: "string" },
+   ·                           ─────────┬────────
+   ·                                    ╰── invalid config definition here
+ 9 │   },
+   ╰────
