+++ cli/tests/ui/config_schema_invalid_required_name.json5
@@ -1,0 +1,10 @@
+{
+  manifest_version: "0.1.0",
+  config_schema: {
+    type: "object",
+    properties: {
+      token: { type: "string" },
+    },
+    required: ["HF_TOKEN"],
+  },
+}
