+++ cli/tests/ui/config_schema_invalid_required_name.stderr
@@ -1,0 +1,15 @@
+Error: manifest::invalid_config_schema
+
+  × check failed
+  ╰─▶ invalid config definition at required[0]: required config property name
+      `HF_TOKEN` is invalid; config property names must start with a lowercase
+      ASCII letter and contain only lowercase ASCII letters, digits, and
+      underscores; double underscores are not allowed. Uppercase letters are
+      not allowed; use `hf_token` instead
+   ╭─[<CARGO_MANIFEST_DIR>/tests/ui/config_schema_invalid_required_name.json5:8:16]
+ 7 │     },
+ 8 │     required: ["HF_TOKEN"],
+   ·                ─────┬────
+   ·                     ╰── invalid config definition here
+ 9 │   },
+   ╰────
