+++ cli/tests/ui/config_schema_invalid_type_keyword.json5
@@ -1,0 +1,11 @@
+{
+  manifest_version: "0.1.0",
+  config_schema: {
+    type: "object",
+    properties: {
+      token: {
+        type: 42,
+      },
+    },
+  },
+}
