+++ cli/tests/ui/config_schema_non_local_ref.stderr
@@ -1,0 +1,12 @@
+Error: manifest::invalid_config_schema
+
+  × check failed
+  ╰─▶ invalid config definition at properties.token.$ref: `$ref` must be local
+      to the same config_schema (`#` or `#/...`)
+    ╭─[<CARGO_MANIFEST_DIR>/tests/ui/config_schema_non_local_ref.json5:9:22]
+  8 │     properties: {
+  9 │       token: { $ref: "#token" },
+    ·                      ────┬───
+    ·                          ╰── invalid config definition here
+ 10 │     },
+    ╰────
