+++ cli/tests/ui/config_schema_properties_wrong_type.stderr
@@ -1,0 +1,12 @@
+Error: manifest::invalid_config_schema
+
+  × check failed
+  ╰─▶ invalid config definition at properties: `properties` must be an object
+      in Amber config_schema, got array
+   ╭─[<CARGO_MANIFEST_DIR>/tests/ui/config_schema_properties_wrong_type.json5:5:17]
+ 4 │     type: "object",
+ 5 │     properties: [],
+   ·                 ─┬
+   ·                  ╰── invalid config definition here
+ 6 │   },
+   ╰────
