+++ cli/tests/ui/config_schema_unknown_keyword.json5
@@ -1,0 +1,7 @@
+{
+  manifest_version: "0.1.0",
+  config_schema: {
+    type: "object",
+    uiLabel: "Root config",
+  },
+}
