+++ cli/tests/ui/config_schema_unknown_keyword.status
@@ -1,0 +1,1 @@
+success
