+++ cli/tests/ui/config_schema_unsupported_keyword.json5
@@ -1,0 +1,9 @@
+{
+  manifest_version: "0.1.0",
+  config_schema: {
+    type: "object",
+    patternProperties: {
+      ".*": { type: "string" },
+    },
+  },
+}
